+++ scalr/provider.go
@@ -101,6 +101,7 @@ func Provider() *schema.Provider {
 			"scalr_run_trigger":                    resourceScalrRunTrigger(),
 			"scalr_service_account":                resourceScalrServiceAccount(),
 			"scalr_service_account_token":          resourceScalrServiceAccountToken(),
+			"scalr_slack_integration":              resourceScalrSlackIntegration(),
 			"scalr_tag":                            resourceScalrTag(),
 			"scalr_variable":                       resourceScalrVariable(),
 			"scalr_vcs_provider":                   resourceScalrVcsProvider(),